requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
librosa>=0.10.0
numpy>=1.24.0
scikit-image>=0.21.0
opencv-python>=4.8.0 